Improve calculations where end date is in the future
If a bet had an end-date, we would always use this to calculate the
total expected amount. This meant that if a bet had the range of
a full year entered already (Jan-Dec) with a value of 100/month,
in March we would expectd their balance to be 1200 (12*100) already
instead of the 300 that it should be. The postgres function now takes
the lesser of current date and end date to do this calculation leading
to more expected results for end users.