Update dependencies to fix CI/CD (#33)
* Update psycopg2 to 2.9.3

This is to ensure to have a version that is compatible
with M1 Macs

* Update Flask to 1.1.4

This is to avoid issues with a non compatible `itsdangerous` dependency
being installed and used. See e.g. this test run:
https://github.com/k-nut/csa-share-management/runs/5257972390?check_suite_focus=true

* Pin Markupsafe verison

* Cache property to fix test

In the prior versionm this would cause the issue explained
at https://sqlalche.me/e/14/bhk3