Set required fields in `Bet` to be non nullable
We just had a situation where one of the endpoints
failed on the live version because we had stored a bet
without a value. Bets without values don't make sense
from a domain model perspective though so we should forbid them.
The same goes for bets withouts share_ids or start_dates so we
set those fields to be non nullable as well.