Add `ptyhon-dateutil` requirement
`dateutil` is used in `solawi/fints_import.py` but was
missing from the `pyproject.toml` file and hence from
the lockfile. When I tried to deploy a new version to Heroku
with the poetry buildpack enabled, the app would crash on
start since this dependency was not available at runtime.