Do not write unset fields as null to database
With the current code, when a user would send a request
where an `Optional` field was not set, pydantic would
return that as `None` to the code and we would then
continue to write this `None`/`null` value to the database.
This means that when users e.g. tried to mark a deposit as
`ignore`, the frontend would send `{"ignore": true}` which
would be turned by pydantic into
`dict(ignore=True, is_security=None)` which would then lead
us to delete the `is_security` value in the db. Adding the
`exclude_unset` parameter changes the behavior in pydantic
so that only the set values are included in the returned dict.
With this change, we now only update the values that are actually
included in the request.